third_party/blink/web_tests/editing/execCommand/insetText-table-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>

test(() => assert_selection(
  [
    `<div contenteditable><table style="display:table-row;">
    <tbody><tr><td>
    <textarea></textarea>^</td></tr></tbody>
    </table>|
    </div>`,
  ],
  'insertText',
  [
    `<div contenteditable><table style="display:table-row;">
    <tbody><tr><td>
    <textarea></textarea>|</td></tr></tbody></table>
    </div>`,
  ]),
  'InsertText should not crash.');
</script>
